+++ src/app/auth/login/login.component.ts
@@ -25,6 +25,12 @@ export class LoginComponent implements OnInit {
     private sessionService: SessionService) { }
 
   ngOnInit() {
+    const hasReloadedUser = sessionStorage.getItem('hasReloadedUser');
+    if (hasReloadedUser == "false" || hasReloadedUser == null) {
+      window.location.reload();
+      sessionStorage.setItem('hasReloadedUser', 'true');
+    }
+
     this.login = new FormGroup({
       email: new FormControl('', [
         Validators.required,
@@ -60,6 +66,10 @@ export class LoginComponent implements OnInit {
         };
       },
     });
+  }
+
+  goToRegister() {
+    sessionStorage.setItem('hasReloadedUser', 'false');
   }
 
   get email() { return this.login.get('email'); }