+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Todas las notas de lanzamiento se registran en este archivo. Este proyecto sigue la especificación [Semantic Versioning](https://semver.org/).
 
+## 3.0.0
+
+### Cambios Generales
+
+- [ABCJ-54] Se ha agregado la funcionalidad de [MOVIL] - Completar perfil candidato. [Ver Detalles](https://abcjobsg16.atlassian.net/browse/ABCJ-54)
+- [ABCJ-51] Se ha agregado la funcionalidad de [MOVIL] - Registrar candidato. [Ver Detalles](https://abcjobsg16.atlassian.net/browse/ABCJ-51)
+
+
 ## 2.2.0
 
 ### Cambios Generales